VMS-35246: Reorder client files, fix coding style and compiler warnings
Renamed and moved to the corresponding place part of the files, related to screen and voice recording.

Fusion was replaced with `nx_reflect` in part of the functionality.

Some clang compilation warnings are fixed on the way:
* using deprecated datetime locale constants
* overloading which hides virtual functions
* using operator `|` or enums of different types

+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/export/dialogs/private/export_settings_dialog_p.cpp
@@ -465,13 +465,13 @@ void ExportSettingsDialog::Private::updateOverlayImage(
 ExportOverlayWidget* ExportSettingsDialog::Private::overlay(ExportOverlayType type)
 {
     const auto index = int(type);
-    return index < m_overlays.size() ? m_overlays[index] : nullptr;
+    return index < (int) m_overlays.size() ? m_overlays[index] : nullptr;
 }
 
 const ExportOverlayWidget* ExportSettingsDialog::Private::overlay(ExportOverlayType type) const
 {
     const auto index = int(type);
-    return index < m_overlays.size() ? m_overlays[index] : nullptr;
+    return index < (int) m_overlays.size() ? m_overlays[index] : nullptr;
 }
 
 void ExportSettingsDialog::Private::setMediaPreviewWidget(nx::vms::client::desktop::AsyncImageWidget* widget)
